feat: refactor Tabs component structure
