+++ src/components/tabs/tabs-content.tsx
@@ -1,0 +1,8 @@
+import * as TabsPrimitive from "@radix-ui/react-tabs"
+import type * as React from "react"
+
+import { cn } from "@/lib/utils"
+
+export function TabsContent({ className, ...props }: React.ComponentProps<typeof TabsPrimitive.Content>) {
+  return <TabsPrimitive.Content data-slot="tabs-content" className={cn("flex-1 outline-none", className)} {...props} />
+}
