+++ src/components/tabs/types.ts
